+++ src/video.hpp
@@ -80,7 +80,7 @@ public:
 	int getx() const;
 	int gety() const;
 
-	std::pair<int, int> get_dpi_scale_factor() const;
+	std::pair<float, float> get_dpi_scale_factor() const;
 
 	//blits a surface with black as alpha
 	void blit_surface(int x, int y, surface surf, SDL_Rect* srcrect=nullptr, SDL_Rect* clip_rect=nullptr);
